refactor: 서버 패키지 구조 분리
