+++ src/backend/search-server/src/main/java/com/bbebig/searchserver/domain/history/domain/DmChatMessage.java
@@ -1,6 +1,6 @@
-package com.bbebig.searchserver.domain.search.domain;
+package com.bbebig.searchserver.domain.history.domain;
 
-import com.bbebig.commonmodule.kafka.dto.*;
+import com.bbebig.commonmodule.kafka.dto.ChatFileDto;
 import jakarta.persistence.Id;
 import lombok.Builder;
 import lombok.Data;
@@ -12,14 +12,12 @@ import java.util.List;
 import static com.bbebig.commonmodule.kafka.dto.ChatMessageDto.*;
 
 @Data
-@Document(collection = "channel_chat_messages")
+@Document(collection = "dm_chat_messages")
 @Builder
-public class ChannelChatMessage {
+public class DmChatMessage {
 
 	@Id
 	private Long id;
-
-	private Long serverId;
 
 	private Long channelId;
 
@@ -28,6 +26,8 @@ public class ChannelChatMessage {
 	private String content;
 
 	private List<ChatFileDto> attachedFiles;
+
+	private List<Long> targetMemberIds;
 
 	private LocalDateTime createdAt;
 
@@ -51,5 +51,3 @@ public class ChannelChatMessage {
 		return this.isDeleted;
 	}
 }
-
-
