+++ src/backend/search-server/src/main/java/com/bbebig/searchserver/domain/history/dto/HistoryDtoConverter.java
@@ -1,9 +1,9 @@
-package com.bbebig.searchserver.domain.search.dto;
+package com.bbebig.searchserver.domain.history.dto;
 
 import com.bbebig.commonmodule.kafka.dto.ChatMessageDto;
-import com.bbebig.searchserver.domain.search.domain.ChannelChatMessage;
+import com.bbebig.searchserver.domain.history.domain.ChannelChatMessage;
 import com.bbebig.searchserver.domain.search.domain.ChannelChatMessageElastic;
-import com.bbebig.searchserver.domain.search.domain.DmChatMessage;
+import com.bbebig.searchserver.domain.history.domain.DmChatMessage;
 import com.bbebig.searchserver.domain.search.domain.DmChatMessageElastic;
 
 import java.util.List;
