+++ src/backend/search-server/src/main/java/com/bbebig/searchserver/domain/search/dto/SearchMessageDtoConverter.java
@@ -1,14 +1,8 @@
 package com.bbebig.searchserver.domain.search.dto;
 
-import com.bbebig.searchserver.domain.search.domain.ChannelChatMessage;
 import com.bbebig.searchserver.domain.search.domain.ChannelChatMessageElastic;
-import com.bbebig.searchserver.domain.search.domain.DmChatMessage;
 import com.bbebig.searchserver.domain.search.domain.DmChatMessageElastic;
-import com.bbebig.searchserver.domain.search.dto.HistoryResponseDto.GetChannelMessageResponseDto;
-import com.bbebig.searchserver.domain.search.dto.HistoryResponseDto.GetDmMessageResponseDto;
 import org.springframework.data.domain.Page;
-
-import java.util.List;
 
 import static com.bbebig.searchserver.domain.search.dto.SearchResponseDto.*;
