+++ src/backend/search-server/src/main/java/com/bbebig/searchserver/domain/search/service/SearchService.java
@@ -6,22 +6,19 @@ import co.elastic.clients.elasticsearch.core.SearchRequest;
 import co.elastic.clients.elasticsearch.core.SearchResponse;
 import co.elastic.clients.elasticsearch.core.search.Hit;
 import co.elastic.clients.json.JsonData;
-import com.bbebig.searchserver.domain.search.domain.ChannelChatMessage;
 import com.bbebig.searchserver.domain.search.domain.ChannelChatMessageElastic;
-import com.bbebig.searchserver.domain.search.domain.DmChatMessage;
 import com.bbebig.searchserver.domain.search.domain.DmChatMessageElastic;
 import com.bbebig.searchserver.domain.search.dto.SearchMessageDtoConverter;
 import com.bbebig.searchserver.domain.search.dto.SearchOption;
 import com.bbebig.searchserver.domain.search.dto.SearchRequestDto.*;
 import com.bbebig.searchserver.domain.search.dto.SearchResponseDto.*;
-import com.bbebig.searchserver.domain.search.repository.ChannelChatMessageRepository;
-import com.bbebig.searchserver.domain.search.repository.DmChatMessageRepository;
+import com.bbebig.searchserver.domain.history.repository.ChannelChatMessageRepository;
+import com.bbebig.searchserver.domain.history.repository.DmChatMessageRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.io.IOException;
