+++ src/backend/search-server/src/main/java/com/bbebig/searchserver/global/repository/DmRedisRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.bbebig.searchserver.domain.search.repository;
+package com.bbebig.searchserver.global.repository;
 
 import com.bbebig.commonmodule.redis.repository.DmRedisRepository;
 import com.bbebig.commonmodule.redis.util.DmRedisKeys;
